coiote-aws-iot-cloud-formation/certificates/requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.5
boto3==1.18.61
crhelper==2.0.10
pyopenssl==21.0.0